Removed copy of CoAP.net now that the CoAP.net package on nuget is up to date.
Made a start on a interactive setup screen for Bastet (in web UI). Accessing the root page either brings up the main page or, if you've not completed setup, shows the setup page. Whether or not you've completed setup is not stored in the database, instead it's dynamically calculated using a db query (see RootModule.SetupProgress constructor).

Modified admin user to not be configured on commandline, but to always be created with username:"Administrator" and password:"password". This could be moved into an externally built strings resource file later for i18n. Naturally this means the --admin and --password commandline switches are gone.

Signed-off-by: Martin Evans <martindevans@gmail.com>